Вынес разметку в отдельные модули
