+++ src/components/point/markdown/get-destination-select-markdown.js
@@ -1,0 +1,5 @@
+export const getDestinationSelectMarkdown = (data) => {
+  return data._destinations.map((item) => {
+    return `<option value="${item.name}"></option>`.trim();
+  }).join(``);
+};
